Scientific Calculator
An application purely developed using Python and its libraries. The application  was developed as a personal project with a motive of learning GUI development using python or tkinter. This application can provide a lot of ease while doing scientific calculation as it can perform various complex operations. 
Also this application can work as both  'Standard Calculator' and 'Scientific Calculator '. The two modes can easily be accessed by navigating to the file option in the menu bar at the leftmost corner of the window